Support /check command that returns proxy health/progress summary
These checks are essential in detecting post-response crashes, unlogged
transactions, and (eventually) other problems that (may be caused but)
cannot be detected using HTTP transactions alone.

The dependency on Perl v5.14+ is unfortunate, but I hope we can get
away with it (without implementing serialization to JSON).

POP /start and /stop commands now return a health check as well to avoid
following up with an explicit /check transaction.