Support running as root
Given a binary choice, Squid should be started as root because deployed
Squids often start as root and because some features require such start.

When overlord.pl starts as root, it needs to "fix" ownership of
directories it creates (e.g., "overlord/"") for Squid to be able to
access them (after Squid switches from root to nobody or some such).
Added code assumes that the parent directory (e.g.,
/usr/local/squid/var/logs) already has the right ownership.