Support --valgrind-suppressions
... instead of hard-coding etc/valgrind.supp

This approach allows the script user to insist on enabling Valgrind
support (instead of accidentally disabling Valgrind testing due to
accidental lack of Valgrind support in the environment).

This approach is arguably more flexible, but we may need to do more
to also support (less important) "use valgrind if possible" use cases.